+++ app/api/auth/[...nextauth]/route.ts
@@ -1,0 +1,2 @@
+import { handlers } from 'auth';
+export const { GET, POST } = handlers;
