+++ app/api/deleteFile/route.ts
