+++ src/components/SignOutButton.tsx
@@ -1,0 +1,5 @@
+import { signOut } from 'next-auth/react';
+
+export function SignOutButton() {
+  return <button onClick={() => signOut()}>Sign Out</button>;
+}
